+++ tower-http/src/compression/future.rs
@@ -1,7 +1,8 @@
 #![allow(unused_imports)]
 
 use super::{body::BodyInner, CompressionBody};
-use crate::compression_utils::{supports_transparent_compression, WrapBody};
+use crate::compression::predicate::Predicate;
+use crate::compression_utils::WrapBody;
 use crate::content_encoding::Encoding;
 use futures_util::ready;
 use http::{header, HeaderMap, HeaderValue, Response};
@@ -18,17 +19,19 @@ pin_project! {
     ///
     /// [`Compression`]: super::Compression
     #[derive(Debug)]
-    pub struct ResponseFuture<F> {
+    pub struct ResponseFuture<F, P> {
         #[pin]
         pub(crate) inner: F,
         pub(crate) encoding: Encoding,
+        pub(crate) predicate: P
     }
 }
 
-impl<F, B, E> Future for ResponseFuture<F>
+impl<F, B, E, P> Future for ResponseFuture<F, P>
 where
     F: Future<Output = Result<Response<B>, E>>,
     B: Body,
+    P: Predicate,
 {
     type Output = Result<Response<CompressionBody<B>>, E>;
 
@@ -36,12 +39,13 @@ where
     fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
         let res = ready!(self.as_mut().project().inner.poll(cx)?);
 
+        // never recompress responses that are already compressed
+        let should_compress = !res.headers().contains_key(header::CONTENT_ENCODING)
+            && self.predicate.should_compress(&res);
+
         let (mut parts, body) = res.into_parts();
 
-        let body = match (
-            supports_transparent_compression(&parts.headers),
-            self.encoding,
-        ) {
+        let body = match (should_compress, self.encoding) {
             // if compression is _not_ support or the client doesn't accept it
             (false, _) | (_, Encoding::Identity) => {
                 return Poll::Ready(Ok(Response::from_parts(